spawn delayed lookup service from network beacon processor
